+++ tests/testthat/test-output-partition.R
@@ -2,11 +2,15 @@ patrick::with_parameters_test_that(
   "partition functions work",
   .cases = {
     lf <- as_polars_lf(mtcars)
+    sink_ipc <- function(...) lf$sink_ipc(..., compression = "zstd")
+    sink_ipc_sorted <- function(...) {
+      lf$sort("am", "cyl")$sink_ipc(..., compression = "zstd")
+    }
 
     tibble::tribble(
       ~.test_name, ~fn, ~reader, ~fn_sorted,
       "sink_csv", lf$sink_csv, pl$read_csv, lf$sort("am", "cyl")$sink_csv,
-      "sink_ipc", lf$sink_ipc, pl$read_ipc, lf$sort("am", "cyl")$sink_ipc,
+      "sink_ipc", sink_ipc, pl$read_ipc, sink_ipc_sorted,
       "sink_ndjson", lf$sink_ndjson, pl$read_ndjson, lf$sort("am", "cyl")$sink_ndjson,
       "sink_parquet", lf$sink_parquet, pl$read_parquet, lf$sort("am", "cyl")$sink_parquet,
     )
@@ -45,6 +49,7 @@ test_that("approximate_bytes_per_file does not support 2^64 - 1 for now", {
   expect_snapshot(
     as_polars_lf(mtcars)$sink_ipc(
       pl$PartitionBy(out_max_size_max, max_rows_per_file = 2^64 - 1),
+      compression = "zstd",
       mkdir = TRUE
     ),
     error = TRUE
@@ -55,11 +60,15 @@ patrick::with_parameters_test_that(
   "deprecated partition functions work",
   .cases = {
     lf <- as_polars_lf(mtcars)
+    sink_ipc <- function(...) lf$sink_ipc(..., compression = "zstd")
+    sink_ipc_sorted <- function(...) {
+      lf$sort("am", "cyl")$sink_ipc(..., compression = "zstd")
+    }
 
     # Only tests sink_ipc for simplicity
     tibble::tribble(
       ~.test_name, ~fn, ~reader, ~fn_sorted,
-      "sink_ipc", lf$sink_ipc, pl$read_ipc, lf$sort("am", "cyl")$sink_ipc,
+      "sink_ipc", sink_ipc, pl$read_ipc, sink_ipc_sorted,
     )
   },
   code = {